requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
pypdf>=6.0.0
python-docx>=1.2.0
google-generativeai==0.8.3
